Fixed suggested fixes and linted all files
