+++ src/js/CheckingScreen.js
@@ -4,22 +4,22 @@ var MenuItem = require('./MenuItem');
 var CheckModule = require('./CheckModule');
 
 module.exports = React.createClass({   // checkecking screen has menu item and check module in it
-  getInitialState: function(){
+  getInitialState: function() {
     return {
-      checkedStatus:"NOT_CHECKED"// its either , retained,replaced,wrong,
-                                 //or not checked
+      checkedStatus: "NOT_CHECKED"// its either , retained,replaced,wrong,
+                                 // or not checked
     };
   },
 
   // changes the status of checked status to the parameter "status"
   changeCheckedStatus: function(status) {
     this.setState({
-      checkedStatus:status
+      checkedStatus: status
     });
   },
 
-  render: function(){
-    return(
+  render: function() {
+    return (
       <div>
           <MenuItem checkedStatus={this.state.checkedStatus} />
           <CheckModule onCheckedStatusChanged={this.changeCheckedStatus} />
