+++ src/js/index.js
@@ -9,7 +9,7 @@
   const menubar = require('./menubar');
   const CommentBox = require('./CommentBox');
   const NextButton = require('./NextButton');
-  const CheckingScreen =  require('./CheckingScreen');
+  const CheckingScreen = require('./CheckingScreen');
 
   var App = {
     init: function() {
